+++ CHANGELOG.md
@@ -47,6 +47,7 @@
 ## 4.62.0 (2025-10-31)
 
 - added: evmChainId values for all EVM chains
+- fixed: (cosmos) Fix max spend
 
 ## 4.61.4 (2025-10-14)